requirements.txt
APScheduler==3.6.3
asgiref==3.8.1
autopep8==1.6.0
boto3==1.35.38
botocore==1.35.38
cachetools==4.2.2
certifi==2024.8.30
charset-normalizer==2.0.12
Django==5.1.1
django-formtools==2.3
django-storages==1.14.4
idna==3.10
isort==5.10.1
jmespath==1.0.1
pillow==10.4.0
pycodestyle==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2024.2
requests==2.26.0
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==1.26.20
